+++ libs/protocol/src/server/p_server_send_packet.c
@@ -20,6 +20,6 @@ bool p_server_send_packet(p_payload_t *payload, const int client_fd,
             break;
         }
     return client
-        && write(client_fd, &payload->packet_type, sizeof(uint8_t)) != -1
+        && write(client_fd, &payload->packet_type, sizeof(uint16_t)) != -1
         && write(client_fd, payload->data, DATA_SIZE) != -1;
 }
